Extracting out configuration of entities for Guid, Json and Concept so that it can be used in more scenarios.
